+++ app/views/layouts/_navbar.html.erb
@@ -5,10 +5,10 @@
         </li>
         <% if !current_user %>
         <li>
-            <%= link_to 'Register', new_user_path %>
+            <%= link_to 'Register', users_new_user_path %>
         </li>
         <li>
-            <%= link_to 'Login', new_session_path %>
+            <%= link_to 'Login', users_new_session_path %>
         </li>
         <% else %>
         <li>
@@ -21,7 +21,7 @@
             <%= link_to 'My Items', items_path %>
         </li>
         <li>
-            <%= link_to 'Logout', sign_out_path, data: { turbo_method: :delete } %>
+            <%= link_to 'Logout', users_sign_out_path, data: { turbo_method: :delete } %>
         </li>
         <% end %>
     </ul>
