+++ config/routes.rb
@@ -10,13 +10,12 @@ Rails.application.routes.draw do
   resources :items
   root 'landing#index', as: 'landing_page'
 
-  get 'registration', action: 'new', controller: 'registration', as: 'new_user'
-  post 'registration', action: 'register', controller: 'registration', as: 'sign_up'
-  get 'sign-in', action: 'new', controller: 'session', as: 'new_session'
-  post 'sign-in', action: 'sign_in', controller: 'session', as: 'sign_in'
-  delete 'sign-out', action: 'sign_out', controller: 'session', as: 'sign_out'
+  namespace :users, path: '/' do
+    get 'registration', to: 'registration#new', as: 'new_user'
+    post 'registration', to: 'registration#register', as: 'sign_up'
+    get 'sign-in', to: 'session#new', as: 'new_session'
+    post 'sign-in', to: 'session#sign_in', as: 'sign_in'
+    delete 'sign-out', to: 'session#sign_out', as: 'sign_out'
+  end
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
-  # Defines the root path route ("/")
-  # root "articles#index"
 end
